requirements.txt
Flask
neo4j
torch
torch-geometric
pymongo
beautifulsoup4
requests
python-dotenv
rapidfuzz
unidecode
groq
langchain_community
bert-score
sacrebleu